Refactor ChannelManager to subpackage channel
